Add support for itsdangerous serializers, and improve JWT processing speed.
This adds support for URLSafeSerializer and URLSafeTimedSerializer from the
itsdangerous package in Python. These are similar to Django and Flask but
derive their keys differently.

Since these appear very similar to JWTs, the Python processors now exit on
decoding when their timestamp field is invalid. This reduces the work we
have to do for JWTs, although it's still higher for Python-based cookies.